# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Store Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for the store
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (from public storefront)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order with WhatsApp link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithWhatsApp"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # AI endpoints
...
  /billing/status:
    get:
      operationId: getBillingStatus
      tags: [billing]
      summary: Get the current user's billing status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Billing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingStatus"
  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Create a Stripe checkout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
  /billing/portal:
    post:
      operationId: createBillingPortal
      tags: [billing]
      summary: Create a Stripe customer portal session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
...
  /growth/share-message:
    get:
      operationId: getShareMessage
      tags: [growth]
      summary: Get WhatsApp share message for the store
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Share message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareMessage"
  /growth/qr-code:
    get:
      operationId: getQrCode
      tags: [growth]
      summary: Get QR code PNG for the store URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR code image
          content:
            image/png:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        selectedVariants:
          type: object
          additionalProperties:
            type: string
      required: [productId, productName, quantity, price]
    Order:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        customerNote:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        deliveryType:
          type: string
          enum: [delivery, pickup]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - items
        - total
        - status
        - createdAt
    CreateOrderBody:
      type: object
      properties:
        storeId:
          type: integer
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        customerNote:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        deliveryType:
          type: string
          enum: [delivery, pickup]
          nullable: true
      required: [storeId, items]
    OrderWithWhatsApp:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        whatsappUrl:
          type: string
      required: [order, whatsappUrl]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
      required: [status]
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        priceMonthly:
          type: number
        maxProducts:
          type: integer
        maxOrdersPerMonth:
          type: integer
        isUnlimited:
          type: boolean
        features:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - displayName
        - priceMonthly
        - maxProducts
        - maxOrdersPerMonth
        - isUnlimited
        - features
    BillingStatus:
      type: object
      properties:
        plan:
          $ref: "#/components/schemas/Plan"
        status:
          type: string
        ordersUsed:
          type: integer
        ordersLimit:
          type: integer
        productsUsed:
          type: integer
        productsLimit:
          type: integer
        usagePercent:
          type: number
        isNearLimit:
          type: boolean
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
      required:
        - plan
        - status
        - ordersUsed
        - ordersLimit
        - productsUsed
        - productsLimit
        - usagePercent
        - isNearLimit
    CreateCheckoutBody:
      type: object
      properties:
        planName:
          type: string
          enum: [pro, business]
      required: [planName]
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
      required: [url]
...
    ShareMessage:
      type: object
      properties:
        message:
          type: string
        whatsappUrl:
          type: string
        storeUrl:
          type: string
      required: [message, whatsappUrl, storeUrl]