# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Store Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  displayName: string;
  priceMonthly: number;
  maxProducts: number;
  maxOrdersPerMonth: number;
  isUnlimited: boolean;
  features: string[];
}