# lib/api-zod/src/generated/types/shareMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Store Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareMessage {
  message: string;
  whatsappUrl: string;
  storeUrl: string;
}